requirements.txt
# Web Framework and API
fastapi>=0.68.0
uvicorn>=0.15.0
flask>=2.0.0
flask-cors>=3.0.0
gunicorn>=20.1.0

# Data Processing and Analysis
numpy>=1.21.0
pandas>=1.3.0

# NLP and Machine Learning
transformers>=4.30.0
torch>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.0.0
spacy>=3.5.0
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.5.0/en_core_web_sm-3.5.0-py3-none-any.whl

# Vector Store and Document Processing
pinecone-client[grpc]>=2.2.0
langchain>=0.1.0
langchain-pinecone>=0.0.1
langchain_community>=0.0.1
langchain_openai>=0.0.1
langchain_experimental>=0.0.1
langchain_groq>=0.0.1
langchain_huggingface>=0.0.1
huggingface_hub>=0.20.0
pypdf>=3.0.0

# Web Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
selenium>=4.15.0
undetected-chromedriver>=3.5.0

# Data Validation
pydantic>=2.0.0

# Testing
pytest>=7.0.0

# UI
streamlit>=1.28.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
tqdm>=4.65.0
jq>=1.6.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
isort>=5.12.0
mypy>=1.0.0